+++ app/src/main/java/io/github/lazyimmortal/sesame/model/normal/base/BaseModel.java
@@ -67,6 +67,10 @@ public class BaseModel extends Model {
     private static final IntegerModelField toastOffsetY = new IntegerModelField("toastOffsetY", "气泡纵向偏移", 0);
     @Getter
     private static final BooleanModelField enableOnGoing = new BooleanModelField("enableOnGoing", "开启状态栏禁删", false);
+    @Getter
+    private static final BooleanModelField closeCaptchaDialog = new BooleanModelField("closeCaptchaDialog", "弹窗验证(自动关闭)", true);
+    @Getter
+    private static final IntegerModelField closeCaptchaDialogDelay = new IntegerModelField("closeCaptchaDialogDelay", "弹窗验证时间(毫秒)", 3000, 100, 30000);
     
     @Override
     public String getName() {
@@ -114,6 +118,8 @@ public class BaseModel extends Model {
         //modelFields.addField(closeCaptchaDialogVPN);
         modelFields.addField(enableOnGoing);
         modelFields.addField(toastOffsetY);
+        modelFields.addField(closeCaptchaDialog);
+        modelFields.addField(closeCaptchaDialogDelay);
         return modelFields;
     }